% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_token}
\alias{check_token}
\title{Check API token}
\usage{
check_token(token)
}
\arguments{
\item{token}{API token to Hlidac statu}
}
\value{
No return value, called for side effects
}
\description{
Check API token
}
