% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_types}
\alias{check_types}
\title{Check if all social media account types are valid}
\usage{
check_types(types)
}
\arguments{
\item{types}{Vector with types of social media accounts}
}
\value{
No return value, called for side effects
}
\description{
Check if all social media account types are valid
}
