% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_type}
\alias{create_type}
\title{Create query string for osoby/social API endpoint}
\usage{
create_type(types)
}
\arguments{
\item{types}{Vector with types of social media accounts}
}
\description{
Create query string for osoby/social API endpoint
}
