% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_contract}
\alias{get_contract}
\title{Get contract details}
\usage{
get_contract(id, token)
}
\arguments{
\item{id}{ID of contract}

\item{token}{Authorization token}
}
\description{
Get contract details
}
