% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_contract_text}
\alias{get_contract_text}
\title{Get contract text}
\usage{
get_contract_text(id, token = Sys.getenv("HLIDAC_TOKEN"))
}
\arguments{
\item{id}{ID of contract}

\item{token}{Authorization token}
}
\value{
Character vector containing text of the contract
}
\description{
Get contract text
}
\examples{
\dontrun{
get_contract_text(id = "1086905", token = "XXXX")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/api/v2/swagger/index}

\url{https://www.hlidacstatu.cz/api/v1/doc}

\url{https://smlouvy.gov.cz/}
}
