% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dataset_metadata}
\alias{get_dataset_metadata}
\title{Get dataset's metadata}
\usage{
get_dataset_metadata(dataset_id, token)
}
\arguments{
\item{dataset_id}{ID of dataset}

\item{token}{Authorization token}
}
\value{
Response from the server containing metadata related to the specified dataset
}
\description{
Get metadata related to specified dataset.
}
\examples{
\dontrun{get_dataset_metadata("ministri", token = "XXXX")}
}
