% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_datasets}
\alias{get_datasets}
\title{Get list of datasets}
\usage{
get_datasets(token = NULL)
}
\arguments{
\item{token}{Authorization token}
}
\value{
Response from the server containing list of all datasets available.
The list contains 3 elements:
- Total: Total number of datasets available (integer)
- Page: Page of the result (integer)
- Results: Data.frame with data concerning the datasets
}
\description{
Get list of datasets available via Hlídač státu API
}
\examples{
\dontrun{
get_datasets(token = "XXXX")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/api/v2/swagger/index}

\url{https://www.hlidacstatu.cz/api/v1/doc}

Other Datasets: 
\code{\link{get_dataset_data}()},
\code{\link{get_dataset_metadata}()},
\code{\link{get_dataset_record_detail}()}
}
\concept{Datasets}
