% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_subsidy}
\alias{get_subsidy}
\title{Get subsidy}
\usage{
get_subsidy(id, token = Sys.getenv("HLIDAC_TOKEN"))
}
\arguments{
\item{id}{Subsidy ID}

\item{token}{Authorization token}
}
\value{
list with details of the subsidy
}
\description{
Get subsidy
}
\examples{
\dontrun{
get_subsidy("deminimis-1000229862")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/swagger/index.html}

\url{https://www.hlidacstatu.cz/api/v1/doc}

Other Subsidies: 
\code{\link{search_subsidies}()}
}
\concept{Subsidies}
