% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_website_detail}
\alias{get_website_detail}
\title{Get website detail}
\usage{
get_website_detail(id, token = Sys.getenv("HLIDAC_TOKEN"))
}
\arguments{
\item{id}{Website id}

\item{token}{Authorization token}
}
\value{
list with 2 elements:
- availability: data on availability of the website
including response time and HTTP status code every minute
- ssl: assessment of the website's SSL server configuration
}
\description{
Get website detail
}
\examples{
\dontrun{
get_website_detail(id = "10107")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/swagger/index.html}

\url{https://www.hlidacstatu.cz/api/v1/doc}
}
