% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handle_error_response}
\alias{handle_error_response}
\title{Handle response other than 200}
\usage{
handle_error_response(response)
}
\arguments{
\item{response}{Response from the server}
}
\value{
No return value, called for side effects
}
\description{
Handle response other than 200
}
