% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{search_person}
\alias{search_person}
\title{Get person details}
\usage{
search_person(query, party, token = NULL)
}
\arguments{
\item{query}{Search query for a person}

\item{party}{Search query for a political party}

\item{token}{Access token}
}
\description{
Get person details
}
\examples{
\dontrun{
search_person(query = "Kalousek", party = "TOP 09", token = "XXXX")
}
}
