% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_company}
\alias{get_company}
\title{Get company details}
\usage{
get_company(company_name, token)
}
\arguments{
\item{company_name}{Name of the company}

\item{token}{Access token}
}
\description{
Get company details
}
