% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_company}
\alias{get_company}
\title{Get company details}
\usage{
get_company(company_name, token = Sys.getenv("HLIDAC_TOKEN"))
}
\arguments{
\item{company_name}{Name of the company}

\item{token}{Access token}
}
\value{
list with 3 elements:
- ico: ID number of the company
- jmeno: Name of the company
- datoveSchranky: Data post box - electronic repository for communication between
public institutions and private companies/people
}
\description{
Get company details
}
\examples{
\dontrun{
get_company("Agrofert")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/swagger/index.html}

\url{https://www.hlidacstatu.cz/api/v1/doc}
}
\concept{Companies}
