% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_contract_text}
\alias{get_contract_text}
\title{Get contract text}
\usage{
get_contract_text(id, token = NULL)
}
\arguments{
\item{id}{ID of contract}

\item{token}{Authorization token}
}
\description{
Get contract text
}
