% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dataset_data}
\alias{get_dataset_data}
\title{Get data from the dataset}
\usage{
get_dataset_data(
  dataset_id,
  query = NULL,
  page = 1,
  sort = NULL,
  desc = NULL,
  token
)
}
\arguments{
\item{dataset_id}{ID of dataset}

\item{query}{Search query}

\item{page}{Number of page}

\item{sort}{Sort by column}

\item{desc}{Descending or ascending sorting}

\item{token}{Authorization token}
}
\description{
Get data from the dataset
}
\examples{
\dontrun{
get_dataset_data("ministri", token = "XXXX")
get_dataset_data("ministri", query = "Zeman", page = 1, token = "XXXX")
}
}
