% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dataset_metadata}
\alias{get_dataset_metadata}
\title{Get dataset's metadata}
\usage{
get_dataset_metadata(dataset_id, token = NULL)
}
\arguments{
\item{dataset_id}{ID of dataset}

\item{token}{Authorization token}
}
\value{
List containing metadata related to the specified dataset
}
\description{
Get metadata related to specified dataset.
}
\examples{
\dontrun{
get_dataset_metadata("ministri", token = "XXXX")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/api/v2/swagger/index}
\url{https://www.hlidacstatu.cz/api/v1/doc}

Other Datasets: 
\code{\link{get_dataset_data}()},
\code{\link{get_dataset_record_detail}()},
\code{\link{get_datasets}()}
}
\concept{Datasets}
