% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dataset_record_detail}
\alias{get_dataset_record_detail}
\title{Get record detail}
\usage{
get_dataset_record_detail(dataset_id, item_id, token)
}
\arguments{
\item{dataset_id}{ID of dataset}

\item{item_id}{ID of record}

\item{token}{Authorization token}
}
\description{
Get record detail
}
