% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dataset_record_detail}
\alias{get_dataset_record_detail}
\title{Get record detail}
\usage{
get_dataset_record_detail(dataset_id, item_id, token = NULL)
}
\arguments{
\item{dataset_id}{ID of dataset}

\item{item_id}{ID of record}

\item{token}{Authorization token}
}
\value{
list of vectors containing data related to the item
}
\description{
Get record detail
}
\examples{
\dontrun{
get_dataset_record_detail("ministri", item_id = 1, token = "XXXX")
}
}
\seealso{
\code{\link{get_dataset_data}}
\url{https://www.hlidacstatu.cz/api/v2/swagger/index}
\url{https://www.hlidacstatu.cz/api/v1/doc}

Other Datasets: 
\code{\link{get_dataset_data}()},
\code{\link{get_dataset_metadata}()},
\code{\link{get_datasets}()}
}
\concept{Datasets}
