% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_datasets}
\alias{get_datasets}
\title{Get list of datasets}
\usage{
get_datasets(token)
}
\arguments{
\item{token}{Authorization token}
}
\value{
Response from the server containing list of all datasets available
}
\description{
Get list of datasets available via Hlídač státu API
}
