% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_datasets}
\alias{get_datasets}
\title{Get list of datasets}
\usage{
get_datasets(token = Sys.getenv("HLIDAC_TOKEN"))
}
\arguments{
\item{token}{Authorization token}
}
\value{
Response from the server containing a list of all datasets available.
The list contains 3 elements:
- total: Total number of datasets available (integer)
- page: Page of the result (integer)
- results: data concerning the datasets (data.frame)
}
\description{
Get list of datasets available via 'Hlídač státu' API
}
\examples{
\dontrun{
get_datasets()
}
}
\seealso{
\url{https://www.hlidacstatu.cz/swagger/index.html}

\url{https://www.hlidacstatu.cz/api/v1/doc}

Other Datasets: 
\code{\link{get_dataset_data}()},
\code{\link{get_dataset_metadata}()},
\code{\link{get_dataset_record_detail}()}
}
\concept{Datasets}
