% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_person}
\alias{get_person}
\title{Get person details}
\usage{
get_person(person_id, token = NULL)
}
\arguments{
\item{person_id}{Person's ID}

\item{token}{Access token}
}
\description{
Get person details
}
\examples{
\dontrun{
get_person("andrej-babis", token = "XXXX")
}
}
