% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_website_detail}
\alias{get_website_detail}
\title{Get website detail}
\usage{
get_website_detail(id, token)
}
\arguments{
\item{id}{Website id}

\item{token}{Authorization token}
}
\description{
Get website detail
}
