% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_website_detail}
\alias{get_website_detail}
\title{Get website detail}
\usage{
get_website_detail(id, token = NULL)
}
\arguments{
\item{id}{Website id}

\item{token}{Authorization token}
}
\value{
list with 2 elements:
- Availability: data on availability of the website
including response time and HTTP status code every minute
- SSL: assessment of the website's SSL server configuration
}
\description{
Get website detail
}
\examples{
\dontrun{
get_website_detail(id = "10107", token = "XXXX")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/api/v2/swagger/index}
\url{https://www.hlidacstatu.cz/api/v1/doc}
}
