% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{search_contracts}
\alias{search_contracts}
\title{Search contract}
\usage{
search_contracts(query, page = 1, sort = 0, token)
}
\arguments{
\item{query}{Search query}

\item{page}{Page of results}

\item{sort}{Ordering of the results}

\item{token}{Access token}
}
\description{
Search contract
}
