% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{search_person}
\alias{search_person}
\title{Search person}
\usage{
search_person(query, page = NULL, token = Sys.getenv("HLIDAC_TOKEN"))
}
\arguments{
\item{query}{Search query for a person, required parameter}

\item{page}{Page number}

\item{token}{Access token}
}
\value{
data.frame with found persons with the following columns:
- jmeno: given name
- prijmeni: surname
- narozeni: date of birth
- nameId: person's ID
- profile: URL of profile at hlidacstatu.cz
- titulPred: academic titles written before person's name
- titulPo: academic titles written after person's name
}
\description{
Search person
}
\examples{
\dontrun{
search_person(query = "Kalousek")
}
}
\seealso{
\url{https://www.hlidacstatu.cz/swagger/index.html}

\url{https://www.hlidacstatu.cz/api/v1/doc}

Other Persons: 
\code{\link{get_person_social}()},
\code{\link{get_person}()},
\code{\link{search_person_by_parameters}()}
}
\concept{Persons}
